add fun to support add visual to exist dashboard
