+++ src/control-plane/backend/lambda/api/service/quicksight/reporting-utils.ts
@@ -19,33 +19,61 @@ import {
   TransformOperation,
   ColumnTag,
   InputColumn,
-  AnalysisSourceEntity,
-  CreateAnalysisCommandOutput,
-  CreateDashboardCommandOutput,
-  TemplateVersionDefinition,
-  DataSetIdentifierDeclaration,
-  DashboardVersionDefinition,
-  AnalysisDefinition,
+  FilterControl, FilterGroup, ParameterDeclaration, Visual, DashboardVersionDefinition, DataSetIdentifierDeclaration
 } from '@aws-sdk/client-quicksight';
 import { v4 as uuidv4 } from 'uuid';
-import { DataSetProps, waitForAnalysisCreateCompleted, waitForDashboardCreateCompleted } from './dashbaord-ln';
+import { DataSetProps } from './dashbaord-ln';
 import { logger } from '../../common/powertools';
+import { RedshiftDataClient } from '@aws-sdk/client-redshift-data';
+import { ClickStreamStore } from '../../store/click-stream-store';
+import { AssumeRoleCommand, STSClient } from '@aws-sdk/client-sts';
+import { getQuickSightSubscribeRegion } from '../../store/aws/quicksight';
+import { PipelineStackType } from '../../common/types';
+import { CPipeline } from '../../model/pipeline';
 
 export interface VisualPorps {
   readonly name: string;
   readonly sheetId: string;
-  readonly visualContent: any;
-  readonly dataSetConfiguration: any;
-  readonly filterControl: any;
-  readonly parameterDeclarations: any[];
-  readonly filterGroup: any;
+  readonly visual: Visual;
+  readonly dataSetIdentifierDeclaration: DataSetIdentifierDeclaration;
+  readonly filterControl: FilterControl;
+  readonly parameterDeclarations: ParameterDeclaration[];
+  readonly filterGroup: FilterGroup;
   readonly eventCount: number;
 }
 
 export interface DashboardAction {
   readonly action: 'ADD' | 'UPDATE' | 'DELETE';
   readonly visuals: VisualPorps[];
-  readonly dashboardDef: string;
+  readonly dashboardDef: DashboardVersionDefinition;
+}
+
+export interface Status {
+  readonly code: number;
+  readonly message?: string;
+}
+
+export interface DashboardCreateParameters {
+  readonly status: Status;
+  readonly redshiftRegion?: string;
+  readonly redshiftDataClient?: RedshiftDataClient;
+  readonly workgroupName?: string;
+  readonly clusterIdentifier?: string;
+  readonly dbUser?: string;
+  readonly isProvisionedRedshift?: boolean;
+  readonly quickSightPricipal?: string;
+  readonly dataApiRole?: string;
+  readonly datasourceArn?: string;
+}
+
+export interface DashboardCreateInputParameters {
+  readonly action: 'PREVIEW' | 'PUBLISH';
+  readonly accountId: string;
+  readonly projectId: string;
+  readonly appId: string;
+  readonly pipelineId: string;
+  readonly stsClient: STSClient;
+  readonly store: ClickStreamStore;
 }
 
 export const funnelVisualColumns: InputColumn[] = [
@@ -175,171 +203,17 @@ export const createDataSet = async (quickSight: QuickSight, awsAccountId: string
   }
 };
 
-export const createAnalysis = async (quickSight: QuickSight, awsAccountId: string, principalArn: string,
-  sourceEntity: AnalysisSourceEntity, analysisName: string)
-: Promise<CreateAnalysisCommandOutput|undefined> => {
-
-  try {
-    const analysisId = uuidv4();
-
-    logger.info('start to create analysis');
-    const analysis = await quickSight.createAnalysis({
-      AwsAccountId: awsAccountId,
-      AnalysisId: analysisId,
-      Name: analysisName,
-      Permissions: [{
-        Principal: principalArn,
-        Actions: [
-          'quicksight:DescribeAnalysis',
-          'quicksight:UpdateAnalysisPermissions',
-          'quicksight:QueryAnalysis',
-          'quicksight:UpdateAnalysis',
-          'quicksight:RestoreAnalysis',
-          'quicksight:DeleteAnalysis',
-          'quicksight:DescribeAnalysisPermissions',
-        ],
-      }],
-
-      SourceEntity: sourceEntity,
-    });
-    await waitForAnalysisCreateCompleted(quickSight, awsAccountId, analysisId);
-    logger.info(`Create analysis finished. Id: ${analysisId}`);
-
-    return analysis;
-
-  } catch (err: any) {
-    logger.error(`Create QuickSight analysis failed due to: ${(err as Error).message}`);
-    throw err;
-  }
-};
-
-export const createDashboard = async (quickSight: QuickSight, awsAccountId: string, principalArn: string,
-  definition: DashboardVersionDefinition, dashboardName: string)
-: Promise<CreateDashboardCommandOutput|undefined> => {
-  try {
-    const dashboardId = uuidv4();
-
-    logger.info('start to create dashboard');
-    const dashboard = await quickSight.createDashboard({
-      AwsAccountId: awsAccountId,
-      DashboardId: dashboardId,
-      Name: dashboardName,
-      Permissions: [{
-        Principal: principalArn,
-        Actions: [
-          'quicksight:DescribeDashboard',
-          'quicksight:ListDashboardVersions',
-          'quicksight:UpdateDashboardPermissions',
-          'quicksight:QueryDashboard',
-          'quicksight:UpdateDashboard',
-          'quicksight:DeleteDashboard',
-          'quicksight:DescribeDashboardPermissions',
-          'quicksight:UpdateDashboardPublishedVersion',
-        ],
-      }],
-
-      Definition: definition,
-
-    });
-    await waitForDashboardCreateCompleted(quickSight, awsAccountId, dashboardId);
-    logger.info(`Create dashboard finished. Id: ${dashboardId}`);
-
-    return dashboard;
-
-  } catch (err: any) {
-    logger.error(`Create QuickSight dashboard failed due to: ${(err as Error).message}`);
-    throw err;
-  }
-};
-
-export const createTemplateFromDefinition = 
-async (quickSight: QuickSight, awsAccountId: string, principalArn: string, dashboardDef: TemplateVersionDefinition)
-: Promise<string|undefined> => {
-
-  const templateId = uuidv4();
-  const template = await quickSight.createTemplate({
+export const getDashboardDefinitionFromArn = async (quickSight: QuickSight, awsAccountId: string, dashboardId: string)
+: Promise<DashboardVersionDefinition|undefined> => {
+  const dashboard = await quickSight.describeDashboardDefinition({
     AwsAccountId: awsAccountId,
-    TemplateId: templateId,
-    Permissions: [{
-      Principal: principalArn,
-      Actions: [
-        'quicksight:CreateTemplate',
-        'quicksight:DescribeTemplate',
-        'quicksight:UpdateTemplate',
-        'quicksight:DeleteTemplate',
-        'quicksight:PassTemplate',
-        'quicksight:UpdateTemplatePermissions',
-        'quicksight:DescribeTemplatePermissions',
-      ],
-    }],
-    Definition: dashboardDef,
+    DashboardId: dashboardId,
   });
 
-  return template.Arn;
+  return dashboard.Definition
 };
 
-export const getDatasetConfigsOfAnalysis = async (quickSight: QuickSight, awsAccountId: string, analysisId: string)
-: Promise<DataSetIdentifierDeclaration[]|undefined> => {
-  const analysisDef = await quickSight.describeAnalysisDefinition({
-    AwsAccountId: awsAccountId,
-    AnalysisId: analysisId,
-  });
-
-  return analysisDef.Definition?.DataSetIdentifierDeclarations;
-};
-
-export const createTemplateFromDashboard =
-async (quickSight: QuickSight, awsAccountId: string, principalArn: string, _dashboardId: string)
-: Promise<string|undefined> => {
-
-  const templateId = uuidv4();
-  const template = await quickSight.createTemplate({
-    AwsAccountId: awsAccountId,
-    TemplateId: templateId,
-    Permissions: [{
-      Principal: principalArn,
-      Actions: [
-        'quicksight:CreateTemplate',
-        'quicksight:DescribeTemplate',
-        'quicksight:UpdateTemplate',
-        'quicksight:DeleteTemplate',
-        'quicksight:PassTemplate',
-        'quicksight:UpdateTemplatePermissions',
-        'quicksight:DescribeTemplatePermissions',
-      ],
-    }],
-    SourceEntity: {
-      SourceAnalysis: {
-        Arn: '',
-        DataSetReferences: [],
-      },
-    },
-  });
-
-  return template.Arn;
-};
-
-export const getTemplateDefinition = async (quickSight: QuickSight, awsAccountId: string, templateId: string)
-: Promise<TemplateVersionDefinition|undefined> => {
-  const templateDef = await quickSight.describeTemplateDefinition({
-    AwsAccountId: awsAccountId,
-    TemplateId: templateId,
-  });
-
-  return templateDef.Definition;
-};
-
-export const getAnalysisDefinitionFromArn = async (quickSight: QuickSight, awsAccountId: string, analysisId: string)
-: Promise<AnalysisDefinition|undefined> => {
-  const analysisDef = await quickSight.describeAnalysisDefinition({
-    AwsAccountId: awsAccountId,
-    AnalysisId: analysisId,
-  });
-
-  return analysisDef.Definition;
-};
-
-export function applyChangeToDashboard(dashboardAction: DashboardAction) : string {
+export function applyChangeToDashboard(dashboardAction: DashboardAction) : DashboardVersionDefinition {
   try {
     if (dashboardAction.action === 'ADD') {
       return addVisuals(dashboardAction.visuals, dashboardAction.dashboardDef);
@@ -351,36 +225,32 @@ export function applyChangeToDashboard(dashboardAction: DashboardAction) : strin
   }
 };
 
-function addVisuals(visuals: VisualPorps[], dashboardDef: string) : string {
-  const dashboard = JSON.parse(dashboardDef);
+function addVisuals(visuals: VisualPorps[], dashboardDef: DashboardVersionDefinition) : DashboardVersionDefinition {
 
   // add visuals to sheet
   for (const visual of visuals) {
     logger.info('start to add visual');
-    const sheet = findElementWithProperyValue(dashboard, 'Sheets', 'SheetId', visual.sheetId);
+
+    const sheet = findElementWithProperyValue(dashboardDef, 'Sheets', 'SheetId', visual.sheetId);
     if ( sheet !== undefined) {
       //add visual to sheet
-      const charts = findElementByPath(sheet, 'Visuals') as Array<any>;
-      logger.info(`charts: ${JSON.stringify(charts)}`);
-      charts.push(visual.visualContent);
+      const charts = sheet.Visuals!;
+      charts.push(visual.visual);
 
       //add dataset configuration
-      const configs = findElementByPath(dashboard, 'DataSetIdentifierDeclarations') as Array<any>;
-      configs.push(visual.dataSetConfiguration);
+      const configs = dashboardDef.DataSetIdentifierDeclarations!;
+      configs.push(visual.dataSetIdentifierDeclaration);
 
       //add filter
-      const controls = findElementByPath(sheet, 'FilterControls') as Array<any>;
-      logger.info(`controls: ${JSON.stringify(controls)}`);
+      const controls = sheet.FilterControls!;
       controls.push(visual.filterControl);
 
       //add parameters
-      const parameters = findElementByPath(dashboard, 'ParameterDeclarations') as Array<any>;
-      logger.info(`parameters: ${JSON.stringify(parameters)}`);
+      const parameters = dashboardDef.ParameterDeclarations!;
       parameters.push(...visual.parameterDeclarations);
 
       //add dataset configuration
-      const fiterGroups = findElementByPath(dashboard, 'FilterGroups') as Array<any>;
-      logger.info(`fiterGroups: ${JSON.stringify(fiterGroups)}`);
+      const fiterGroups = dashboardDef.FilterGroups!
       fiterGroups.push(visual.filterGroup);
 
       // visual layout
@@ -391,7 +261,6 @@ function addVisuals(visuals: VisualPorps[], dashboardDef: string) : string {
 
       if (elements.length > 0) {
         const lastElement = elements.at(elements.length - 1);
-        logger.info(`lastElement: ${JSON.stringify(lastElement)}`);
         layoutControl.RowIndex = lastElement.RowIndex + lastElement.RowSpan;
         visualControl.RowIndex = lastElement.RowIndex + lastElement.RowSpan + layoutControl.RowSpan;
       }
@@ -399,16 +268,279 @@ function addVisuals(visuals: VisualPorps[], dashboardDef: string) : string {
       layoutControl.ElementId = firstObj.FilterControlId;
       visualControl.RowSpan = visual.eventCount * 2;
 
-      logger.info(`visual.visualContent: ${visual.visualContent}`);
-      visualControl.ElementId = findFirstChild(visual.visualContent).VisualId;
+      visualControl.ElementId = findFirstChild(visual.visual).VisualId;
       elements.push(layoutControl);
       elements.push(visualControl);
 
     }
   }
 
-  return JSON.stringify(dashboard);
+  return dashboardDef;
 };
+
+export async function getDashboardCreateParameters(input: DashboardCreateInputParameters) : Promise<DashboardCreateParameters> {
+
+  //get requied parameters from ddb and stack output.
+  const pipeline = await input.store.getPipeline(input.projectId, input.pipelineId);
+  if (!pipeline) {
+    return { status: { code: 404, message: 'Pipeline not found' } };
+  }
+  const redshiftRegion = pipeline.region;
+
+  logger.info(`redshiftRegion: ${redshiftRegion}`);
+
+  if (!pipeline.dataModeling?.redshift) {
+    return { status: { code: 404, message: 'Redshift not enabled in the pipeline' } };
+  }
+
+  const isProvisionedRedshift = pipeline.dataModeling?.redshift?.provisioned ? true : false;
+  const clusterIdentifier = isProvisionedRedshift ? pipeline.dataModeling?.redshift.provisioned?.clusterIdentifier : undefined;
+  const dbUser= isProvisionedRedshift ? pipeline.dataModeling?.redshift.provisioned?.dbUser : undefined;
+
+  let workgroupName = undefined;
+  let dataApiRole = undefined;
+
+  logger.info(`pipeline: ${pipeline}`);
+
+  const cPipeline = new CPipeline(pipeline);
+  const modelingStackOutputs = await cPipeline.getStackOutputs(PipelineStackType.DATA_MODELING_REDSHIFT);
+
+  console.log(`stackoutput: ${JSON.stringify(modelingStackOutputs)}`);
+  for (const [name, value] of modelingStackOutputs) {
+    if (name.endsWith('WorkgroupName')) {
+      workgroupName = value;
+    }
+    if (name.endsWith('DataApiRoleArn')) {
+      dataApiRole = value;
+    }
+  }
+  if (!workgroupName && !isProvisionedRedshift) {
+    return { status: { code: 404, message: 'Redshift serverless workgroup not found' } };
+  }
+  if (!dataApiRole) {
+    dataApiRole = 'arn:aws:iam::451426793911:role/Clickstream-DataModelingR-RedshiftServerelssWorkgr-1B641805YKFF7';
+    // return {code: 404, message: 'Redshift data api role not found'};
+  }
+
+  let datasourceArn = undefined;
+  let quicksightInternalUser = undefined;
+  const reportingStackOutputs = await cPipeline.getStackOutputs(PipelineStackType.REPORTING);
+  for (const [name, value] of reportingStackOutputs) {
+    if (name.endsWith('DataSourceArn')) {
+      datasourceArn = value;
+    }
+    if (name.endsWith('QuickSightInternalUser')) {
+      quicksightInternalUser = value;
+    }
+  }
+  if (!datasourceArn) {
+    datasourceArn = 'arn:aws:quicksight:us-east-1:451426793911:datasource/clickstream_datasource_project11_50543d10';
+    // return res.status(404).send(new ApiFail('QuickSight data source arn not found'));
+  }
+  if (!quicksightInternalUser) {
+    quicksightInternalUser = 'Admin/liwmin-Isengard';
+    // return res.status(404).send(new ApiFail('QuickSight internal user not found'));
+  }
+
+  //quicksight user name
+  const quicksightPublicUser = pipeline.reporting?.quickSight?.user;
+  if (!quicksightPublicUser) {
+    return { status: { code: 404, message: 'QuickSight user not found' } };
+  }
+
+  let quicksightUser;
+  if (input.action === 'PREVIEW') {
+    quicksightUser = quicksightInternalUser!;
+  } else {
+    quicksightUser = quicksightPublicUser;
+  }
+
+  quicksightUser = 'Admin/liwmin-Isengard';
+
+  const awsPartition = 'aws';
+  const quickSightSubscribeRegion = await getQuickSightSubscribeRegion();
+  logger.info(`quickSightSubscribeRegion: ${quickSightSubscribeRegion}`);
+
+  const quickSightPricipal = `arn:${awsPartition}:quicksight:${quickSightSubscribeRegion}:${input.accountId}:user/default/${quicksightUser}`;
+  console.log(`quickSightPricipal: ${quickSightPricipal}`);
+
+  //get redshift client
+  const credentials = await getCredentialsFromRole(input.stsClient, dataApiRole);
+  const redshiftDataClient = new RedshiftDataClient({
+    region: redshiftRegion,
+    credentials: {
+      accessKeyId: credentials?.AccessKeyId!,
+      secretAccessKey: credentials?.SecretAccessKey!,
+      sessionToken: credentials?.SessionToken,
+    },
+  });
+
+  return {
+    status: { code: 200 },
+    redshiftRegion,
+    redshiftDataClient,
+    quickSightPricipal,
+    dataApiRole,
+    clusterIdentifier,
+    dbUser,
+    workgroupName,
+    datasourceArn,
+    isProvisionedRedshift,
+  };
+
+}
+
+export async function getCredentialsFromRole(stsClient: STSClient, roleArn: string) {
+  try {
+    const assumeRoleCommand = new AssumeRoleCommand({
+      RoleArn: roleArn,
+      RoleSessionName: 'redshift-data-api-role',
+    });
+
+    const response = await stsClient.send(assumeRoleCommand);
+    const credentials = response.Credentials;
+
+    return credentials;
+  } catch (error) {
+    console.error('Error occurred while assuming role:', error);
+    throw error;
+  }
+}
+
+export function getVisualDef(visualId: string, viewName: string) : Visual {
+
+  const visualDef = JSON.parse(readFileSync(join(__dirname, './templates/funnel-chart.json')).toString()) as Visual;
+  const eventNameFiledId = uuidv4();
+  const idFilefId = uuidv4();
+  visualDef.FunnelChartVisual!.VisualId = visualId;
+
+  const fieldWells = visualDef.FunnelChartVisual!.ChartConfiguration!.FieldWells!;
+  const sortConfiguration = visualDef.FunnelChartVisual!.ChartConfiguration!.SortConfiguration!;
+  const tooltipFields = visualDef.FunnelChartVisual?.ChartConfiguration?.Tooltip!.FieldBasedTooltip!.TooltipFields!;
+
+  fieldWells.FunnelChartAggregatedFieldWells!.Category![0].CategoricalDimensionField!.FieldId = eventNameFiledId;
+  fieldWells.FunnelChartAggregatedFieldWells!.Values![0].CategoricalMeasureField!.FieldId = idFilefId;
+  fieldWells.FunnelChartAggregatedFieldWells!.Category![0].CategoricalDimensionField!.Column!.DataSetIdentifier = viewName;
+  fieldWells.FunnelChartAggregatedFieldWells!.Values![0].CategoricalMeasureField!.Column!.DataSetIdentifier = viewName;
+  sortConfiguration.CategorySort![0].FieldSort!.FieldId = idFilefId;
+  sortConfiguration.CategorySort![1].FieldSort!.FieldId = eventNameFiledId;
+
+  tooltipFields[0].FieldTooltipItem!.FieldId = idFilefId;
+  tooltipFields[1].FieldTooltipItem!.FieldId = eventNameFiledId;
+
+  return visualDef;
+}
+
+export interface VisualRalatedDefParams {
+  readonly filterContrl: FilterControl;
+  readonly parameterDeclarations: ParameterDeclaration[];
+  readonly filterGroup: FilterGroup;
+}
+
+export interface VisualRalatedDefProps {
+  readonly timeScopeType: 'FIXED' | 'RELATIVE';
+  readonly sheetId: string;
+  readonly visualId: string;
+  readonly viewName: string;
+  readonly lastN?: number;
+  readonly timeUnit?: string;
+  readonly timeStart?: string;
+  readonly timeEnd?: string;
+}
+
+export function getVisualRalatedDefs(props: VisualRalatedDefProps) : VisualRalatedDefParams {
+
+  const filterControlId = uuidv4();
+  const sourceFilterId = uuidv4();
+  const parameterSuffix = uuidv4().replace(/-/g, '');
+
+  let filterContrl: FilterControl;
+  const parameterDeclarations = [];
+  let filterGroup: FilterGroup;
+
+  if (props.timeScopeType === 'FIXED') {
+    filterContrl = JSON.parse(readFileSync(join(__dirname, './templates/filter-control-datetime.json')).toString()) as FilterControl;
+    filterContrl.DateTimePicker!.FilterControlId = filterControlId;
+    filterContrl.DateTimePicker!.Title = 'event_date between';
+    filterContrl.DateTimePicker!.SourceFilterId = sourceFilterId;
+
+    // const parameterDeclarationStart = JSON.parse(readFileSync(join(__dirname, './templates/datetime-parameter.json')).toString()) as ParameterDeclaration;
+    // parameterDeclarationStart.DateTimeParameterDeclaration!.Name = `dateStart${parameterSuffix}`;
+    // parameterDeclarationStart.DateTimeParameterDeclaration!.TimeGranularity = 'DAY';
+    // parameterDeclarationStart.DateTimeParameterDeclaration!.DefaultValues!.StaticValues = [new Date(props.timeStart!)];
+    // parameterDeclarationStart.DateTimeParameterDeclaration!.DefaultValues!.RollingDate = undefined;
+    // parameterDeclarations.push(parameterDeclarationStart);
+
+    // const parameterDeclarationEnd = JSON.parse(readFileSync(join(__dirname, './templates/datetime-parameter.json')).toString()) as ParameterDeclaration;
+    // parameterDeclarationEnd.DateTimeParameterDeclaration!.Name = `dateEnd${parameterSuffix}`;
+    // parameterDeclarationEnd.DateTimeParameterDeclaration!.TimeGranularity = 'DAY';
+    // parameterDeclarationEnd.DateTimeParameterDeclaration!.DefaultValues!.StaticValues = [new Date(props.timeEnd!)];
+    // parameterDeclarationEnd.DateTimeParameterDeclaration!.DefaultValues!.RollingDate = undefined;
+    // parameterDeclarations.push(parameterDeclarationEnd);
+
+    filterGroup = JSON.parse(readFileSync(join(__dirname, './templates/filter-group.json')).toString()) as FilterGroup;
+
+    filterGroup.FilterGroupId = uuidv4();
+    filterGroup.Filters![0].TimeRangeFilter!.FilterId = sourceFilterId;
+    filterGroup.Filters![0].TimeRangeFilter!.Column!.ColumnName = 'event_date';
+    filterGroup.Filters![0].TimeRangeFilter!.Column!.DataSetIdentifier = props.viewName;
+
+    filterGroup.Filters![0].TimeRangeFilter!.RangeMinimumValue!.StaticValue = new Date(props.timeStart!);
+    filterGroup.Filters![0].TimeRangeFilter!.RangeMaximumValue!.StaticValue = new Date(props.timeEnd!);
+
+    filterGroup.ScopeConfiguration!.SelectedSheets!.SheetVisualScopingConfigurations![0].SheetId = props.sheetId;
+    filterGroup.ScopeConfiguration!.SelectedSheets!.SheetVisualScopingConfigurations![0].VisualIds = [props.visualId];
+
+  } else {
+    filterContrl = JSON.parse(readFileSync(join(__dirname, './templates/filter-control-relative-datetime.json')).toString()) as FilterControl;
+    filterContrl.RelativeDateTime!.FilterControlId = filterControlId;
+    filterContrl.RelativeDateTime!.Title = 'event_date';
+    filterContrl.RelativeDateTime!.SourceFilterId = sourceFilterId;
+
+    const parameterDeclarationStart = JSON.parse(readFileSync(join(__dirname, './templates/datetime-parameter.json')).toString()) as ParameterDeclaration;
+    parameterDeclarationStart.DateTimeParameterDeclaration!.Name = `dateStart${parameterSuffix}`;
+    parameterDeclarationStart.DateTimeParameterDeclaration!.TimeGranularity = 'DAY';
+    parameterDeclarationStart.DateTimeParameterDeclaration!.DefaultValues!.RollingDate!.Expression = `addDateTime(-${props.lastN}, '${props.timeUnit}', truncDate('${props.timeUnit}', now()))`;
+    parameterDeclarationStart.DateTimeParameterDeclaration!.DefaultValues!.StaticValues = undefined;
+    parameterDeclarations.push(parameterDeclarationStart);
+
+    const parameterDeclarationEnd = JSON.parse(readFileSync(join(__dirname, './templates/datetime-parameter.json')).toString()) as ParameterDeclaration;
+    parameterDeclarationEnd.DateTimeParameterDeclaration!.Name = `dateEnd${parameterSuffix}`;
+    parameterDeclarationEnd.DateTimeParameterDeclaration!.TimeGranularity = 'DAY';
+    parameterDeclarationEnd.DateTimeParameterDeclaration!.DefaultValues!.RollingDate!.Expression = 'addDateTime(1, \'DD\', truncDate(\'DD\', now()))';
+    parameterDeclarationEnd.DateTimeParameterDeclaration!.DefaultValues!.StaticValues = undefined;
+    parameterDeclarations.push(parameterDeclarationEnd);
+
+    let unit = 'DAY';
+    if (props.timeUnit == 'WK') {
+      unit = 'WEEK';
+    } else if (props.timeUnit == 'MM') {
+      unit = 'MONTH';
+    } else if (props.timeUnit == 'Q') {
+      unit = 'QUARTER';
+    }
+
+    filterGroup = JSON.parse(readFileSync(join(__dirname, './templates/filter-group-relative.json')).toString()) as FilterGroup;
+
+    filterGroup.FilterGroupId = uuidv4();
+    filterGroup.Filters![0].RelativeDatesFilter!.FilterId = sourceFilterId;
+    filterGroup.Filters![0].RelativeDatesFilter!.Column!.ColumnName = 'event_date';
+    filterGroup.Filters![0].RelativeDatesFilter!.Column!.DataSetIdentifier = props.viewName;
+
+    filterGroup.Filters![0].RelativeDatesFilter!.RelativeDateValue = props.lastN;
+    filterGroup.Filters![0].RelativeDatesFilter!.TimeGranularity = unit;
+
+    filterGroup.ScopeConfiguration!.SelectedSheets!.SheetVisualScopingConfigurations![0].SheetId = props.sheetId;
+    filterGroup.ScopeConfiguration!.SelectedSheets!.SheetVisualScopingConfigurations![0].VisualIds = [props.visualId];
+
+  }
+
+  return {
+    parameterDeclarations,
+    filterContrl,
+    filterGroup,
+  };
+}
 
 function findElementByPath(jsonData: any, path: string): any {
   const pathKeys = path.split('.');
@@ -474,3 +606,4 @@ function findElementWithProperyValue(root: any, path: string, property: string, 
     return undefined;
   }
 }
+
