+++ src/control-plane/backend/lambda/api/service/quicksight/sql-builder.ts
