+++ src/control-plane/backend/lambda/api/service/quicksight/templates/dashboard.json
@@ -3,7 +3,7 @@
   "Sheets": [
       {
           "SheetId": "",
-          "Name": "sheet1",
+          "Name": "",
           "FilterControls": [],
           "Visuals": [],
           "Layouts": [
